Moved DEFAULT_CONFIG from global scope to Config
DEFAUL_CONFIG_JSON represented json file that should be used in case the
program could not find any config file in executable directory.
It was in global scope sa in case of exception it could not be caught.
Now its stored inside Config class which fits and can be caught also.